Decouple JavaFX frontend from Chatot logic
