+++ src/main/java/chatot/Launcher.java
@@ -4,6 +4,6 @@ import javafx.application.Application;
 
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Chatot.class, args);
+        Application.launch(Main.class, args);
     }
 }
